# frontend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Ecommerce Playground API",
    "description": "API para sistema de e-commerce de farmacia (Arquitectura de Servicios Refactorizada v2.1)",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/images/products/{product_id}/upload": {
      "post": {
        "tags": [
          "Images"
        ],
        "summary": "Upload Product Image",
        "description": "Orquesta la subida de una imagen para un producto:\n1. Procesa y guarda el archivo f\u00edsico (usando ImageService).\n2. Actualiza la 'image_url' del producto en la DB (usando ProductService).",
        "operationId": "upload_product_image_api_images_products__product_id__upload_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "product_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Product Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_upload_product_image_api_images_products__product_id__upload_post"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Product"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/images/products/{product_id}/image": {
      "delete": {
        "tags": [
          "Images"
        ],
        "summary": "Delete Product Image",
        "description": "Orquesta la eliminaci\u00f3n de la imagen de un producto:\n1. Obtiene la URL de la imagen (usando ProductService).\n2. Borra el archivo f\u00edsico (usando ImageService).\n3. Pone la 'image_url' a NULL en la DB (usando ProductService).",
        "operationId": "delete_product_image_api_images_products__product_id__image_delete",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "product_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Product Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Product"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/images/business/logo": {
      "post": {
        "tags": [
          "Images"
        ],
        "summary": "Upload Business Logo",
        "description": "Orquesta la subida del logo del negocio:\n1. Procesa y guarda el archivo (usando ImageService).\n2. Actualiza la 'logo_url' en la DB (usando BusinessService).",
        "operationId": "upload_business_logo_api_images_business_logo_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_upload_business_logo_api_images_business_logo_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BusinessInfo"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/images/business/icon": {
      "post": {
        "tags": [
          "Images"
        ],
        "summary": "Upload Business Icon",
        "description": "Orquesta la subida del icono del negocio:\n1. Procesa y guarda el archivo (usando ImageService).\n2. Actualiza la 'icon_url' en la DB (usando BusinessService).",
        "operationId": "upload_business_icon_api_images_business_icon_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_upload_business_icon_api_images_business_icon_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BusinessInfo"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/images/business/banner": {
      "post": {
        "tags": [
          "Images"
        ],
        "summary": "Upload Business Banner",
        "description": "Orquesta la subida del banner del negocio:\n1. Procesa y guarda el archivo (usando ImageService).\n2. Actualiza la 'banner_url' en la DB (usando BusinessService).",
        "operationId": "upload_business_banner_api_images_business_banner_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_upload_business_banner_api_images_business_banner_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BusinessInfo"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/images/products/{product_id}/gallery": {
      "get": {
        "tags": [
          "Images"
        ],
        "summary": "Get Product Gallery",
        "description": "Obtiene todas las im\u00e1genes de la galer\u00eda de un producto.\nEndpoint p\u00fablico para mostrar en ProductDetailModal.",
        "operationId": "get_product_gallery_api_images_products__product_id__gallery_get",
        "parameters": [
          {
            "name": "product_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Product Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "additionalProperties": true
                  },
                  "title": "Response Get Product Gallery Api Images Products  Product Id  Gallery Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Images"
        ],
        "summary": "Upload To Product Gallery",
        "description": "A\u00f1ade una nueva imagen a la galer\u00eda del producto.\nSi is_main=True, esta imagen se convierte en la principal.",
        "operationId": "upload_to_product_gallery_api_images_products__product_id__gallery_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "product_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Product Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_upload_to_product_gallery_api_images_products__product_id__gallery_post"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Upload To Product Gallery Api Images Products  Product Id  Gallery Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/images/products/{product_id}/gallery/{image_id}/set-main": {
      "put": {
        "tags": [
          "Images"
        ],
        "summary": "Set Main Gallery Image",
        "description": "Establece una imagen de la galer\u00eda como la imagen principal del producto.",
        "operationId": "set_main_gallery_image_api_images_products__product_id__gallery__image_id__set_main_put",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "product_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Product Id"
            }
          },
          {
            "name": "image_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Image Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Set Main Gallery Image Api Images Products  Product Id  Gallery  Image Id  Set Main Put"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/images/products/{product_id}/gallery/{image_id}": {
      "delete": {
        "tags": [
          "Images"
        ],
        "summary": "Delete Gallery Image",
        "description": "Elimina una imagen de la galer\u00eda del producto.",
        "operationId": "delete_gallery_image_api_images_products__product_id__gallery__image_id__delete",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "product_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Product Id"
            }
          },
          {
            "name": "image_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Image Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Delete Gallery Image Api Images Products  Product Id  Gallery  Image Id  Delete"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Body_upload_business_icon_api_images_business_icon_post": {
        "properties": {
          "file": {
            "type": "string",
            "format": "binary",
            "title": "File"
          }
        },
        "type": "object",
        "required": [
          "file"
        ],
        "title": "Body_upload_business_icon_api_images_business_icon_post"
      },
      "Body_upload_business_logo_api_images_business_logo_post": {
        "properties": {
          "file": {
            "type": "string",
            "format": "binary",
            "title": "File"
          }
        },
        "type": "object",
        "required": [
          "file"
        ],
        "title": "Body_upload_business_logo_api_images_business_logo_post"
      },
...
      "Body_upload_product_image_api_images_products__product_id__upload_post": {
        "properties": {
          "file": {
            "type": "string",
            "format": "binary",
            "title": "File"
          }
        },
        "type": "object",
        "required": [
          "file"
        ],
        "title": "Body_upload_product_image_api_images_products__product_id__upload_post"
      },
...
      "BusinessInfo": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name",
            "default": "E-Commerce"
          },
          "rif": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rif"
          },
          "address": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Address"
          },
          "phone": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Phone"
          },
          "contact": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Contact"
          },
          "social_networks": {
            "items": {
              "$ref": "#/components/schemas/SocialNetwork"
            },
            "type": "array",
            "title": "Social Networks"
          },
          "logo_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Logo Url"
          },
          "icon_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Icon Url"
          },
          "banner_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Banner Url"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "type": "object",
        "title": "BusinessInfo",
        "description": "Informaci\u00f3n del negocio."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "Product": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          },
          "name": {
            "type": "string",
            "maxLength": 200,
            "minLength": 1,
            "title": "Name"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "sku": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sku"
          },
          "price": {
            "type": "string",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Price"
          },
          "stock": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Stock",
            "default": 0
          },
          "category": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Category"
          },
          "image_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Image Url"
          },
          "is_featured": {
            "type": "boolean",
            "title": "Is Featured",
            "default": false
          },
          "is_discount": {
            "type": "boolean",
            "title": "Is Discount",
            "default": false
          },
          "discount_percentage": {
            "type": "string",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Discount Percentage",
            "default": 0.0
          },
          "banner_assignment": {
            "type": "string",
            "title": "Banner Assignment",
            "default": "main"
          },
          "final_price": {
            "type": "string",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Final Price",
            "description": "Precio final calculado con descuento aplicado.",
            "readOnly": true
          }
        },
        "type": "object",
        "required": [
          "name",
          "price",
          "final_price"
        ],
        "title": "Product"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}